TCES-10 Refactored components
